66 errors!

+++ src/convertType.m
@@ -1,6 +1,6 @@
 function typeddata = convertType(x)
 if ~testToolboxes('Parallel Computing Toolbox')
-    typeddata = single(x);
+    typeddata = double(x);
 else
     typeddata = gpuArray(single(x));
 end
